requirements.txt
fastapi
uvicorn
loguru
python-multipart
psycopg2-binary
passlib
bcrypt
python-jose
PyJWT
email-validator
